V4L: Add driver for s5k6a3 image sensor
This patch adds subdev driver for Samsung S5K6A3 raw image sensor.
As it is intended at the moment to be used only with the Exynos
FIMC-IS (camera ISP) subsystem it is pretty minimal subdev driver.
It doesn't do any I2C communication since the sensor is controlled
by the ISP and its own firmware.
This driver can be updated in future, should anyone need it to be
a regular subdev driver where the main CPU communicates with the
sensor directly.

Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>
---
Changes since v1:
 - added missing pm_runtime_disable(),
 - removed subdev name overriding,
 - s/S5K6A3_DEF_PIX/S5K6A3_DEFAULT_,
 - merged patch adding v4l2-async API support,
 - clock-frequency property is now optional and a default frequency
   value will be used when it is missing, rather than bailing out,
 - reset GPIO made mandatory as it is required for proper power
   on/off sequences and all known boards use it,
 - maximum image size is now 1412x1412 pixels, as specified in
   in the S5K6A3YX datasheet.
 - regulator_enable()/disable() used instead of the regulator bulk
   API to ensure proper regulators enable/disable sequences.