exynos-gsc: Remove GSC_{SRC, DST}_FMT flags
The GSC_SRC_FMT, GSC_DST_FMT flags are currently set in VIDIOC_S_FMT ioctl
and cleared in VIDIOC_REQBUFS(0). In between the flags are used to figure out
if scaling ratio check need to be performed. This an incorrect behaviour as
it should be assumed there is always a valid configuration on a video device.
Fix this by removing those flags and also remove an unused 'frame' local
variable.

Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>