ARM: EXYNOS: Enable power domains for !CONFIG_GENERIC_POWER_DOMAINS
When CONFIG_GENERIC_POWER_DOMAINS is not set enable required power
domain when driver of a device that belongs to the domain is bound.
This allows devices within the power domain to work also when
PM_SLEEP and PM_RUNTIME is not set.

The power domains code is always compiled in now for ARCH_EXYNOS.

Thanks-to: Tomasz Figa <t.figa@samsung.com>
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>